refactor useOnrampeventlistener
